+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.5.0 - TBD
+
+* Optimization: push restrictions over rename & autosummarize.
+
 # 0.4.1 - March 9st, 2018
 
 * Fix Image#restrict optimization that pushed down restrictions on right attributes